+++ lib/smart_health_cards_test_kit/health_card.rb
@@ -1,4 +1,4 @@
-module SmartHealthCards
+module SmartHealthCardsTestKit
   module HealthCard
     def payload_from_jws(jws)
       return nil unless jws.present?
