+++ lib/smart_health_cards_test_kit/shc_payload_verification.rb
@@ -1,4 +1,4 @@
-module SmartHealthCards
+module SmartHealthCardsTestKit
   class SHCPayloadVerification < Inferno::Test
     include HealthCard
 
@@ -40,7 +40,7 @@ module SmartHealthCards
       decompressed_payload_array = []
 
       credential_strings.split(',').each do |credential|
-        jws = SmartHealthCards::JWS.from_jws(credential)
+        jws = SmartHealthCardsTestKit::Utils::JWS.from_jws(credential)
         payload = payload_from_jws(jws)
 
         warning do
