+++ lib/smart_health_cards_test_kit/version.rb
@@ -1,0 +1,3 @@
+module SmartHealthCardsTestKit
+  VERSION = '0.1.0'
+end
