+++ spec/smart_health_cards/shc_fhir_validation_spec.rb
@@ -1,4 +1,4 @@
-RSpec.describe SmartHealthCards::SHCFHIRValidation do
+RSpec.describe SmartHealthCardsTestKit::SHCFHIRValidation do
   let(:suite) { Inferno::Repositories::TestSuites.new.find('smart_health_cards_test_suite') }
   let(:session_data_repo) { Inferno::Repositories::SessionData.new }
   let(:runner) { Inferno::TestRunner.new(test_session: test_session, test_run: test_run) }
